refactor(cli): extract session.py keybindings to _session_keybindings
Moves the prompt_toolkit key-binding rules (Escape / Ctrl-C / Left /
Right tab nav) and the _NS_STATE module-level namespace tracker out
of amx/cli_support/session.py into a focused
amx/cli_support/_session_keybindings.py.

_NS_STATE moves with the keybindings because they are its primary
read/writer. session.py keeps line 1361's write via the re-export, so
no behavior changes.

session.py: 1628 -> 1577 LOC (-51).
